Fix description box glitching.
The description box was glitching between regular and mobile view
when switching between variables with/without url's in their
descriptions. This was due to the row <div> not having a proper
bootstrap column class.

The urls would still overrun into the map, so they where changed
to hyper links with only the domain as the desplay text.